Move sonar drivers from drv_mb1242.(h|cpp) to mb1242.(h|cpp)
